requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
sentence-transformers>=2.5.0
faiss-cpu==1.7.4
numpy>=1.24.3
requests==2.31.0
Werkzeug==3.0.1
